ParticleUtil - add option for bukkitColor - also move bukkitColor from display to other
